CR Comments, extranting KDF
